Fixed warings! Happy business. And some location of files in ModelConvert

+++ blib/SpriteBatch.h
@@ -53,7 +53,7 @@ namespace blib
 		glm::mat4 matrix;
 		FBO* fbo;
 
-		int cacheStart;
+        size_t cacheStart;
 		VBO* vbo;
